refactoring all api

+++ ProductHub/Controllers/ProductController.cs
@@ -12,117 +12,111 @@ namespace ProductHub.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class ProductController : ControllerBase
+    public class ProductController(IMapper mapper, IProductService productService, IImageFileService imageService) : ControllerBase
     {
-        private readonly IProductService _productService;
-        private readonly IImageFileService _imageService;
-        private readonly IMapper _mapper;
-        public ProductController(IMapper mapper, IProductService productService, IImageFileService imageService)
-        {
-            _productService = productService;
-            _mapper = mapper;
-            _imageService = imageService;
-        }
+        private readonly IProductService _productService = productService;
+        private readonly IImageFileService _imageFileService = imageService;
+        private readonly IMapper _mapper = mapper;
 
         [Route("{id}")]
         [HttpGet]
         public async Task<ActionResult<ProductDto>> GetProduct(int id)
         {
-            var product = await _productService.Get(id);
-            var productDto = _mapper.Map<ProductDto>(product);
+            var product = await _productService.GetById(id);
 
-            return Ok(productDto);
+            if (product is null)
+                return NotFound("Product Not Found");
+
+            return Ok(_mapper.Map<ProductDto>(product));
         }
 
         [HttpGet]
-        public async Task<ActionResult<ProductDto>> GetProducts()
+        public async Task<ActionResult<ProductsDto>> GetProducts()
         {
-            var products = await _productService.Get();
+            var products = await _productService.GetAll();
 
             if (products is null)
-            {
                 return NotFound("Products Not Found");
-            }
 
-            var productDto = _mapper.Map<List<ProductDto>>(products);
-
-            return Ok(productDto);
+            return Ok(_mapper.Map<List<ProductsDto>>(products));
         }
 
         [Route("category/{idCategory}")]
         [HttpGet]
-        public async Task<ActionResult<ProductDto>> GetProductsByCategory(int idCategory)
+        public async Task<ActionResult<ProductsDto>> GetProductsByCategory(int idCategory)
         {
             var products = await _productService.GetProductsByCategory(idCategory);
 
             if (products is null)
-            {
                 return NotFound("Products Not Found");
-            }
 
-            var productDto = _mapper.Map<List<ProductDto>>(products);
-
-            return Ok(productDto);
+            return Ok(_mapper.Map<List<ProductsDto>>(products));
         }
 
         [HttpPut]
         public async Task<ActionResult<ProductDto>> UpdateProduct([FromBody] UpdateProductDto updateProductDto)
         {
             var product = _mapper.Map<Product>(updateProductDto);
-            var _product = await _productService.Update(product);
+            var existingProduct = await _productService.Update(product);
 
-            if (_product is null)
-                return NotFound("Product not found.");
+            if (existingProduct is null)
+                return NotFound("Product Not Found");
 
-            var ProductDto = _mapper.Map<ProductDto>(_product);
-
-            return Ok(ProductDto);
+            return Ok(_mapper.Map<ProductDto>(existingProduct));
         }
 
         [HttpPost]
-        public async Task<ActionResult<ProductDto>> AddProduct([FromBody] CreateProductDto createProductDto)
+        public async Task<ActionResult> AddProduct([FromBody] CreateProductDto createProductDto)
         {
             var product = _mapper.Map<Product>(createProductDto);
-            var _product = await _productService.Create(product);
+            var existingProduct = await _productService.Create(product);
 
-            var ProductDto = _mapper.Map<ProductDto>(_product);
+            if (existingProduct is null)
+                return BadRequest("Product was not created");
 
-            return Ok(ProductDto);
+            return Ok("Product Successfully Added");
         }
 
         [HttpPost]
         [Route("comments")]
-        public async Task<ActionResult<ProductDto>> AddCommentProduct([FromBody] CreateCommentDto createCommenttDto)
+        public async Task<ActionResult> AddCommentProduct([FromBody] CreateCommentDto createCommenttDto)
         {
             var comment = _mapper.Map<Comment>(createCommenttDto);
-            var _product = await _productService.AddCommentToProduct(comment);
+            var product = await _productService.AddCommentToProduct(comment);
 
-            var ProductDto = _mapper.Map<ProductDto>(_product);
+            if (product is null)
+                return BadRequest("Product was not created");
 
-            return Ok(ProductDto);
+            return Ok("Comment Successfully Added");
         }
 
         [HttpPost]
         [Route("images")]
-        public async Task<ActionResult<ProductDto>> AddImagesProduct([FromForm] FileUpload file)
+        public async Task<ActionResult> AddImagesProduct([FromForm] FileUpload file)
         {
-            if (file == null || file.files.Length == 0)
+            if (file == null || file.Files.Length == 0)
             {
                 return BadRequest("Invalid file.");
             }
 
-            var fileName = Guid.NewGuid().ToString() + Path.GetExtension(file.files.FileName);
+            var fileName = Guid.NewGuid().ToString() + Path.GetExtension(file.Files.FileName);
 
-            using var stream = file.files.OpenReadStream();
-            var result = await _imageService.Upload(fileName, stream);
+            using var stream = file.Files.OpenReadStream();
+            var resultImageFileOperation = await _imageFileService.Upload(fileName, stream);
 
-            CreateImageDto createImageDto = new() { Url = result.FileName, ProductId = file.productId };
+            if (!resultImageFileOperation.ProcessedSuccessfully)
+                return BadRequest($"The file was not uploaded: {resultImageFileOperation.Message}");
+
+            CreateImageDto createImageDto = new() { Url = resultImageFileOperation.FilePath, ProductId = file.ProductId };
 
             var image = _mapper.Map<Image>(createImageDto);
 
-            var r = await _productService.AddImagesToProduct([image]);
+            var resultProductSerive = await _productService.AddImagesToProduct([image]);
 
-            return Ok(new { FilePath = result });
+            if (resultProductSerive is null)
+                return BadRequest("Image was not created");
+
+            return Ok("Image Successfully Added");
         }
 
         [HttpDelete]
@@ -130,26 +124,35 @@ namespace ProductHub.Controllers
         public async Task<ActionResult> DeleteCommentFromProduct([FromBody] DeleteCommentDto deleteCommentDto)
         {
             var comment = _mapper.Map<Comment>(deleteCommentDto);
-            var response = await _productService.DeleteCommentToProduct(comment);
-            return Ok(response);
+            var product = await _productService.DeleteCommentToProduct(comment);
+
+            if (product is null)
+                return BadRequest("Comment was not deleted");
+
+            return Ok("Comment Successfully Deleted");
         }
 
         [HttpDelete]
         [Route("Image")]
         public async Task<ActionResult> DeleteImageFromProduct([FromBody] DeleteImageDto deleteImageDto)
         {
-            try
-            {
-                var imageResult = await _productService.GetImageFromProduct(deleteImageDto.ProductId, deleteImageDto.Id);
-                var result = _imageService.Delete(imageResult.Url);
-                var image = _mapper.Map<Image>(deleteImageDto);
-                var response = await _productService.DeleteImageToProduct(image);
-                return Ok(response);
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, $"Error: {ex.Message}");
-            }
+            var imageResult = await _productService.GetImageFromProduct(deleteImageDto.ProductId, deleteImageDto.Id);
+
+            if (imageResult is null)
+                return NotFound("Image Not Found");
+
+            var resultImageFileOperation = await _imageFileService.Delete(imageResult.Url);
+
+            if (!resultImageFileOperation.ProcessedSuccessfully)
+                return BadRequest($"The file was not deleted: {resultImageFileOperation.Message}");
+
+            var image = _mapper.Map<Image>(deleteImageDto);
+            var product = await _productService.DeleteImageToProduct(image);
+
+            if (product is null)
+                return BadRequest("Image was not deleted");
+
+            return Ok("Image Successfully Deleted");
         }
     }
 }
